LayoutTests/storage/indexeddb/unprefix-workers-expected.txt

[Worker] Check that IDBFactory is available through the prefixed or unprefixed entry point.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/unprefix.js
PASS [Worker] String(self.indexedDB) is "[object IDBFactory]"
PASS [Worker] String(self.webkitIndexedDB) is "[object IDBFactory]"
PASS [Worker] IDBCursor is non-null.
PASS [Worker] IDBDatabase is non-null.
PASS [Worker] IDBFactory is non-null.
PASS [Worker] IDBIndex is non-null.
PASS [Worker] IDBKeyRange is non-null.
PASS [Worker] IDBObjectStore is non-null.
PASS [Worker] IDBRequest is non-null.
PASS [Worker] IDBTransaction is non-null.
PASS successfullyParsed is true

TEST COMPLETE
